Remove non-top level imports, cobbler configurable, dict changes.
Per PR review I removed the non top-level imports, I made the
cobbler URL configurable (with default to our URL). Instead of
using if's to use the various arch names I made a new function
that instead gets the info from a single dict which is easier to
manage to add new entries to.

I also added test cases for the new function that uses the dict.