ceph_manager: update ec_pool creation parameters
As of https://github.com/ceph/ceph/pull/1475 the erasure code parameters
are controled via the osd set erasure_code_profile command instead of
being inlined and duplicated in the crush ruleset creation and the pool
creation. There is no need to create the crush ruleset, it is done
implicitly.

Signed-off-by: Loic Dachary <loic@dachary.org>