integration tests for teuthology-lock --list --all
Verify that teuthology-lock --list --all

  * outputs valid JSON
  * contains a string which is a clue the JSON is what we want
  * the number of items returned matches what the paddles
    server provides.

Forge a teuthology.yaml configuration file that uses the paddles server
available from the integration environment instead of the default.

The lock.py main function is modified to accept a namespace instead of
using sys.argv. It makes it consistent with the other commands and
allows the test to forge an argument list without modifying sys.argv.

Signed-off-by: Loic Dachary <loic@dachary.org>