Introduces a 'docker' task that leverages maestro-ng
Task that allows to execute orchestrate docker containers from
teuthology. Supported maestro-ng commands are 'start', 'stop' and
'pull'. Tests for this task are also present. A new dependency is added
to setup.py (maestro-ng).

Besides the 'command' option, the docker task has two options that
specify what services are being operated on: 'services' and 'roles'. The
first one is used to specify services in maestro-ng's format. The latter
signals the creation of services for the entries in teuthology's 'roles'
configuration.