transparent OpenStack provisioning for teuthology-suite
The teuthology-openstack command is a wrapper around teuthology-suite
that transparently creates the teuthology cluster using OpenStack
virtual machine.

For machines of machine_type == openstack in paddles, when locking a
machine, an instance is created in the matching OpenStack cluster with:

   openstack server create redhat

And renamed into redhat010 if assigned the IP x.x.x.10/24. It is then
locked in paddles which has been prepare with one slot for each
available IP in the range.

An OpenStack cluster is defined in the .teuthology.yaml file as follows:

openstack:
  user-data: teuthology/openstack-{os_type}-{os_version}-user-data.txt
  nameserver: 167.114.252.136
  machine:
    disk: 10 # GB
    ram: 7000 # MB
    cpus: 1
  volumes:
    count: 0
    size: 1 # GB
  ssh-key: /home/ubuntu/.ssh/id_rsa
  clusters:
    ovh:
      flavor-select-regexp: ^vps-ssd
      openrc.sh: teuthology/test/integration/openrc.sh
      subnet: 167.114.224.0/19
      images:
        ubuntu-14.04: Ubuntu 14.04
        centos-7.0: Centos 7

When the machine is unlocked, it is destroyed with

   openstack server delete redhat010

The python-openstackclient command line is used instead of the
corresponding API because it is well maintained and documented.

Integration tests require an OpenStack tenant.

Signed-off-by: Loic Dachary <loic@dachary.org>