suite: use --subset X/Y as hints
When the number of jobs cannot be subdivided by Y while retaining the
property of having all facets in each set of jobs, it is incorrect to
display the X/Y subset of the jobs. Instead of subdividing the sets
according to X/Y, use Y as mincyclicity but use the cyclicity() from
the matrix as the maximum number of subsets.

http://tracker.ceph.com/issues/13855 Fixes: #13855

Signed-off-by: Loic Dachary <loic@dachary.org>