suite: --subset X/Y may inflate the matrix when Y is large
If Y is larger than the size of the larger matrix found at the first
recursion step, the resulting matrix will be inflated to match Y.
This is not incorrect but it makes it more difficult to understand
what is going on from the developer point of view.

Cap Y to the largest mincyclicity of the matrices included in a Sum().

Signed-off-by: Loic Dachary <loic@dachary.org>