misc: Don't exclude valid block devices in get_block_devices()
The initial filtering to major 8 & 254 is clearly not enough.

By reading the Documentation/devices.txt file from the kernel, it appears that
the /proc/partitions can be full of many devices types.

This patch implements a selection of items we are interested in. We remove all
devices known we are not interested in and keep the remaining.

The benefit of that patch is not considering the device name but only its type.
Prior this patch, cciss, xen, sdaz and many others fun devices or configurations
were totally ignored as they didn't matched the [s:v]d? scheme.

A new test is added to test the sda[a-z] scheme too.

Signed-off-by: Erwan Velu <erwan@redhat.com>