user-data: start salt-minion with teuthology VM as master
This commit installs salt-minion and points it to the teuthology VM as salt-master.

When the teuthology VM is provisioned, it gets a dnsmasq instance pre-populated
with hostname records for all IP addresses in all available subnets. One of
these corresponds to the teuthology VM itself. The target VMs always have
"nameserver $TEUTHOLOGY_VM_IP_ADDR" in their /etc/resolv.conf, so they are
capable of resolving these teuthology-openstack-specific hostnames.

Signed-off-by: Nathan Cutler <ncutler@suse.com>