salt: implement ping_minions_* tasks (serial and parallel, in wait loop)
Use job_id instead of run name and use a job_id grain to avoid pinging minions
from any other test jobs that might be running.

Typically we will try to ping minions as the first thing we do after starting
salt-minion.service. Under such conditions it is not reasonable to assume the
minions will be reliably pingable on the first try.

Signed-off-by: Jan Fajerski <jfajerski@suse.com>
Signed-off-by: Nathan Cutler <ncutler@suse.com>