openstack: upload absolute paths to remote teuthology machine
If any absolute paths are given on the teuthology-openstack command line,
assume they are local YAML files and upload them to the remote teuthology
machine before running teuthology-suite.

If the local YAML file (absolute path) is PATH, the remote
path will be /home/ubuntu/yaml/$PATH. The remote file is clobbered if
it already exists.

Fixes: https://github.com/SUSE/teuthology/issues/46
Fixes: https://github.com/SUSE/teuthology/issues/56
Signed-off-by: Nathan Cutler <ncutler@suse.com>