misc: add flexibility to delete_ and move_file
Add check argument to delete_file to provide the ability to ignore a
failed delete (say if the file to delete does not exist)
Add preserver_perms argument to move_file to make the preservation of
permissions optional. This makes move_file usable if the target
file does not yet exist.

Signed-off-by: Jan Fajerski <jfajerski@suse.com>