images: support SLE-12-SP[123]
This commit adds user-data files/symlinks where missing so we
can utilize our new SLE-12-SP[123] images when they are available.

Since the new images will have the SDK built-in, we omit the SUSEConnect calls
from the user-data.

Since the status of Salt in SLE-12-SP1 is unclear and we don't ship DeepSea for
any SES version that runs on SP1, I omitted the Salt enablement for SP1.

Signed-off-by: Nathan Cutler <ncutler@suse.com>