openstack: refrain from installing RPMs in sle-12.3 user-data
The teuthology-sle-12.3-x86_64 image no longer has any repos
(for better compatibility with our CI, which uses test-repo feature
to add artifacts), so any packages we need for teuthology to work
must be pre-installed in the images.

In other words, there's no point in trying to install packages
here, because without repos it's doomed to fail.

Signed-off-by: Nathan Cutler <ncutler@suse.com>