install: drop purge_data and related code
Not doing this step during the teardown phase of the install task makes it
easier to implement alternative deployment tasks (i.e. tests that deploy Ceph
clusters, but do not call ceph.py).

While the install task causes the ceph-osd package to be installed and creates
the OSD directories under /var/lib/ceph, it is actually the ceph task that
populates these directories so in any case their "unpopulation" should be
implemented there, and not here in the install task.

In the final analysis, however, this code is no longer needed now that all test
nodes (virtual and bare metal) are reimaged between runs.

Signed-off-by: Nathan Cutler <ncutler@suse.com>