teuthology/misc: do not fail when pull_directory(a_dir_does_not_exist)
there is chance that /var/log/ceph (and other directories) is not
created when cephadm fails, but we always try to collect /var/log/ceph
by tar the directory on the test node and transfer it back for archiving
it. when we fail to do so, an exception is thrown, and this exception is
in turn logged as the cause of the failure instead of the one which
fails the test.

in this change, all `tarfile.TarError`s are caught in
`misc.pull_directory()`, as normally, we don't consider it a fatal
failure when `misc.pull_directory()` fails because of `tar`.

two types of failures are caught here. we could use `tarfile.TarError`
here, as both of them inherit from this class, but it'd be easier to
understand to be explicit. as only these two types are expected.

Signed-off-by: Kefu Chai <kchai@redhat.com>