worker: send job config as a file and setup logging
Dispatcher was initially using stdin to send job config to worker
using popen.communicate() which caused the dispatcher to wait for
the worker to finish executing. To avoid that this commit saves
the job config in a temp file and sends the file descriptor to the
worker to read the job config from. It also sets up logging for
workers inside the job's suite logging directory.

Signed-off-by: Shraddha Agrawal <shraddha.agrawal000@gmail.com>