don't try to delete job from server if first or last in suite
This commit removes trying to delete job from server in case the
job is first or last in suite. As first and last job are not
reported to the server in the first place, we don't need to delete
them.

Signed-off-by: Shraddha Agrawal <shraddha.agrawal000@gmail.com>
(cherry picked from commit 6365d8d8e0b28acdc04db23ea9da8dba9c1e1362)