packaging: add 'install_per_major_version' option support
In case for the distros having the same major version, such as for
RHEL 8/8.1/8.2, etc, and there has only one repo using the major
version in the url for one specified package.

The git builder class will just try to get the os version from the
remote node and then uses it to build the request uri. This patch
will make the git build classh discard the minor version if the
'install_per_major_version' is specified.

Signed-off-by: Xiubo Li <xiubli@redhat.com>