orchestra/remote: do not log stdout when calling Remote.read_file()
We do not want to dump everything to the log the `read_file()` method
produce, for example, the target file contents, especially if it is binary.
It is distracting the user and it is better to just log which file it is
trying to read for debugging purpose.

so, in this change, Remote.read_file() and Remote.write_file() are
reworked to use SFTP client instead of "dd ... of=/dev/stdout".

Signed-off-by: Kefu Chai <kchai@redhat.com>