exceptions: change error message for CommandFailedError
For the convenience of user, print the command that led to
CommandFailedError in error message as a str than as a list or a tuple.
This makes it more readable and enables the user to copy and use the
command without any hassles.

This commit also uses the opportunity to rename the variable "command"
to "cmd" and "self.command" to "self.cmd" in CommandFailedError for
convenience and so that it's easy to keep statements under 80
characters.

Signed-off-by: Rishabh Dave <ridave@redhat.com>