orchestra/remote: channge default value for a mktemp() param
Empty string is a much better default value, compared to None which is
the current value, for remote.RemoteShell.mktemp()'s parameter data
because when data is set to None even the creation of an empty file is
skipped. To avoid this, set the default value to empty string and raise
an exception for the case when data type of parameter data is not str.

Signed-off-by: Rishabh Dave <ridave@redhat.com>