orchestra/remote: throw exception on reconnect failed
In run method if a reconnect required there is a remote.reconnect
method called which is supposed to return True or False.
On failure it is the best to raise an exception early,
to make it clear what was the original error instead of
waiting until somewhere later teuthology fails with
some undeclared variables.

Signed-off-by: Kyr Shatskyy <kyrylo.shatskyy@suse.com>