qa: correct line processing logic
Most runs don't use --no-nested-subset and for those that used --subset,
the if conditions would correctly pickup "seed" (when it mattered).
However, when --subset was not specified in the original run, the "seed"
was not correctly picked up. Therefore, the result of inserting the "if
no_nested_subset is None:" before the "elif seed is None:" caused it to
never read the seed for most folks teuthology runs.

Fixes: https://tracker.ceph.com/issues/55815
Signed-off-by: Patrick Donnelly <pdonnell@redhat.com>