teuthology: fix the problem that git_build_url defaults to github Signed-off-by: Sun Junnan <sunjunnan_yewu@cmss.chinamobile.com>
The build_git_url function creates a github connection by default when the project name does not belong to one of ceph,ceph-cm-ansible,ceph-qa-suite.
Modify the function to achieve that if project name is not one of the above three, create the ceph_git_base_url connection in the teutology.yaml file.