teuthology/suite: use paddles solely for rerun info
teuthology.front no longer allows access to the run directories for general
users. In fact, it's not required because we are already getting run
information from paddles.

This commit makes it more stringent in --rerun handling: if
seed/subset/no_nested_subset do not match, exit failure rather than try to
continue. Without a subset, many suites are so large that teuthology-suite will
run for several minutes before ultimately failing to schedule due to the
quantity of jobs. It also doesn't make sense to rerun a suite when these values
do not match.

Signed-off-by: Patrick Donnelly <pdonnell@redhat.com>