orchestra/remote: add resolve_ip method
Add utility method to resolve a hostname from within remote host.
This is useful to resolve ip address of the remote host itself,
because getting ip address from transport object of ssh is not
suitable because it may have only bastion host and port, which
is not relevant for the purpose of configuring a cluster.

Signed-off-by: Kyr Shatskyy <kyrylo.shatskyy@clyso.com>