lock.ops.unlock_one_safe: Invert run-match logic
When unlock_one_safe is called with run_name, the caller means to express
"unlock this node if it belongs to this run".
When it is called with run_name and job_id, it means "unlock this node if it
belongs to this job in this run".
We had inverted the logic, causing leaks on reimage failures.

Signed-off-by: Zack Cerza <zack@cerza.org>
(cherry picked from commit b9a0b03d807de6204e2f01e3622993856097cde7)