dispatcher: do not put targets in orig config
There is introduced targets in orig.config.yaml in the dispatcher code,
however it should not be here so the original config should be still
usable with teuthology (teuthology-run) with arbitrary targets yaml
file, which is provided separately.

Also we don't need to change job status to running when trying to lock
nodes. We change the status when actually trying to start supervisor.

Signed-off-by: Kyr Shatskyy <kyrylo.shatskyy@clyso.com>